my-tutor-flow/my-tutor-flow-agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
anthropic>=0.34.0
python-dotenv==1.0.1
